gui: move internal bitcoind config to bitcoind module
As part of this change, the struct fields were made public.
